Further improvement on collision handling
Improves issue #9

If a tile coliision occurs, the ball is not imediately moved out of the
tile, but the correct direction is reflected. If the ball collided with
the same tile in the previous frame, the collision will now be ignore,
    giving the ball ennough time to move fully out of the tile.

Corner collisions are not handled, and more testing needs to be done.