Improved text rendering
Added error checking for font loading ( will print an error message if
font couldn't be loaded )

Added a new function for making text into a SDL_Surface. The function
takes the same arguments as the original SDL TTF one (
TTF_RenderText_Solid() ) but will in addition to calling
it also call SetDisplayFormat() making sure the SDL_Surface is optimized
for bliting. The fucntion also checks if the SDL_Surface was succesfuly
created and print an error message if not.