Improved paddle hit reflection
Fixes #2

The ball is now reflected in a horizontal direction depending on where
it hit on paddle. The ammount horizontal movement is
increasing/decreaing linearly. The vertical direction is the same as
previous, just oposite. The direction is normalized after hitting paddle
to prevent unintentional speed incrase.

Addded protection to prevent a collision from being detected two times
in a row.