Change in gmae mode is now handled in Run()
Changes in game mode are now handled by MenuManager. MenuManager has a
bool that states wether the game state has changed. When this boos is
read from GameManager, this value will be set back to false, meaning it
will only be true the first time you check this boolean variable.

This means MenuManager handles everything related to the current state
and switching between them, which is better than it being handled by
both GameManager and MenuManager.