Added TCPConnection and NetManager
TCPConnection is a simple class build on top of SDL2_nets functions.
It has functionality for both acting as both client and server ( server
not tested. )

NetManager is intended to manage TCPConnection. There will be one
TCPConnection actie at startup. This one will be kept alive till the
game shuts down. It will connect the game to a swerer. The second one wil be a P2P connection to the oponent this can be both server and client.