Center and flip board
Implements issue #20 but further improvement on the issue might be
needed.

Board will now be centered for both players. It will also be flipped for
the non-server player. When a board is removed at the local player, the
verticaly oposite tile is removed for the remote player. I.e the local
player hits and detrsoys the top-center tile, the bottom-center is
destroyed in the board of the remote player. These tiles will be
identical expcept y-position.