Fixed bug causing balls collisions to look wrong on oposite instance
Balls from the server instance will appear to be colidion wrong on the
client server ( and oposite. ) This is because the ball has oposite
direction and the board is flipped.

This means that if it collides with the bottom site of the ball on the
server side, it will be colliding with the top side on client side.