Fixed issue making local board having wrong scale some times
The local player sometimes got the message with the new board scale from
client before changing to a new board. This results in the scale being
(potentially) wrong.

Fixed by making server dictate when the board is changed. This should
work becaue both player will always know when a tile is removed.

TODO:
-Make the mssage also contain new name of the board.
-Rename message to "NewBoard" or something similar