Improved loading of conifg
-ConfigLoader is now GameManager class member.
-LoadConfig() is execute just once, in Init()
-Added config for Ball speed in fast mode, whether or not fast mode is
active, Bullet speed  and chance of getting a bonus.

Note;
The chance of bonus is noted 1/x, where x is the number in the config
file. A higher number means a lower chance.

Todo:
Change the Bonus chance number in config so that a higher number means
a higher chance. Or, alternatively, two numbers x / y.