Fixed Connection startup error
How to reproduce:
-Start a new game on one client, joing that game on the other.
-This MAY cause the second player to get stuck on the multiplayer name
while the server displays the player name with '7' appended to it.

Fix:
-In added a space to << operator for PlayerName message in TCPMessage.

Info:
-SDL may send several messages as one long string with no space between
the different messages.
-When sending a PlayerName message, the << operator didn't add a space
at the end so when the message GameStateChanged ( type 7 ) was sent
after, the message type was just appended to the string that holds the
playyer name.
-The next message would then get the message type ID from the next
message's object ID. The object ID for this message isn't used, so it's
always set to 0. Message Type 0 is PaddlePosition, preventing the
message from being reported as an unkown message type.