Fixed issues with high speed Bullets
How to reproduce:
-Set Bullet speed to a high value in config ( for instance 9.0 )
-Start a game of any type
-Fire a Bullet
-The Bullet might pass through seceral Tiles before hitting and
destroying a tile, leving the other Tiles unharmed.

Fix:
-Added a function "HasHitTile" in Bullet that takes a Tile as argument
and checks for intersection. It will return true as long as the top of the
Bullet is higher that the bottom of the Tile, meaning it will return true for
Tiles the Bullet has passed completely through.
-Extended GameManager::UpdateBullets() so that it stores the lowest Tile
a Bullet has hit and handles the intersection for this when all of the
Tiles has been checked. This means that the lowest/first Tile will
always be the one that's removed.
-If SuperBullet is acrive, the Tile will be removed as long as it
intersects according to the HasHitTile function.