Fixed issue when loading new Board
-DeleteDeadBullets() will be called at the end of UpdateBullets() this
is because the bounds check will delete Bullets, which won't get cleaned
if this function is not run.
-GenerateBoard() is now run when by server when client has connected.
-Level done will only be checked when a Tile has beed killed, this
elimiated error when a level starts with destrouable Tiles ( destroyalbe tiles aren't counted when counting remainng Tiles )
-Edited GameManager::AddTile() to take an ID, this is because client
gets Tiles from server and the ID needs to be the same on both sides.
-Edited PhysicsManager::CreateTile() so that if tileID == -1, ( this happens only when CreateTile() is called on server, and a new, unique ID is assigned. On the client, this function is called with the ID from the message recieved.