Improved particle generation
-Added function GetCenter() and CenterInRect(). GetPosition() returns
the center of the Rect and CenterInRect() places a smaller rect inside
the sender of a larger rect and returns the top-left position of this
secon rect. This is used to place particles in the iddle of a Tile.
-Renderer::GeneratePartiles() now takes a Tile as arguent. It places the
particles in the middle of this Tile using CenterInTile(). It also gives
the particle the same texture as the tile that was hit.
-For now, only regular Tiles triggers particles.

TODO:
-Make all Tiles generate particles.
-Improve the performance of the particle system ( SetAlphaMod()? )
-Implement OpenGL for use with particles.