Fixed issue with two player not spawning Tiles on client
How to reporudce
-Start two player game
-Sometimes the Tiles will not spawn on remote

Cause
-On remote, the Tiles where recieved before Reset() was called. The
Reset() would then delete all the Tiles!

Fix
-Chnaged WasGameStarted() so that it will only trigger when prev
GameState was not InGameWait, Pause or Lobby. Meaning it will only
trigger when a game is started from main menu ( single player game )
-Moved calls to Reset() to JoinGame() this method is called just before
JoinGame message is sent.
-GenerateBoard() will be called first when server recieves JoinGame
message. And since, on the client side, Reset() is called before sending
of JoinGame, Reset() is guaranteed to be called before the Board is
generated.

Notes:
-Also includes som changes to HandleGameStateChange()
-Added WasGameResumed in MenuManager than simply tells where the game
was in Pause and is now InGame