Added ConfigList
-ConfigList now handles all options
-Works like GameList, but in this case, the list does not contain the
information, ony the elements. This is because the items can be changed
from within the list ( which ia MenuManager terriroty ) The ConfigList
class is only for viweing and scrolling through the itmes.

Implementation:
-Added ConfigList to Rendered ( rendering, intialization and adding
elemets to id )
-Added MoveRight() to ConfigItem. Will move all elemtes of the
ConfigItem right ( pluss/min, value, caption )
-Added SetRectX() and SetRectY() to ConfigItems. Works by calculating
how much it needs to be moved, then uses MoveRight()/MoveDown() to move
all elements.
-Added the ConfigList itself, which has the AddConfigItem() that takes a
ConfigItem ptr and the ConfigValueType, adds it to the internal map of
the class and moves it to the appropriate pos in the list
-Renamed height to itemsTop in List because it's the pos of the next top
item and not the height of the actual list.